search_engine/requirements.txt
# Django
Django
drf-yasg
drf-spectacular[sidecar]
djangorestframework

meilisearch
requests
beautifulsoup4
nltk

scikit-learn
pandas
lxml

python-dotenv